Enhance sales order processing and schema updates
- Cast `TubeWeight` and `TotalProductionTime` to `decimal` in `ProductionAllotmentController.cs` for type consistency.
- Add `MarkSalesOrderItemWebAsProcessed` method in `SalesOrderWebController.cs`.
- Introduce `IsProcess` and `ProcessDate` fields in `SalesOrderWebDTOs.cs` for tracking processing status.
- Update `ISalesOrderWebService.cs` with `MarkSalesOrderItemWebAsProcessedAsync` method.
- Modify `ApplicationDbContextModelSnapshot.cs` to reflect schema changes, including nullable fields and new processing fields.
- Update `ProductionAllotment.cs` model to make fields nullable and adjust types.
- Enhance `SalesOrderWeb.cs` model with processing status and audit fields.
- Implement new processing method in `SalesOrderWebService.cs`.
- Change database connection string in `appsettings.Development.json` to `localhost`.
- Add migration `20251129050203_add_colDate` for schema updates.

+++ Controllers/StorageCaptureController.cs
@@ -188,7 +188,7 @@ namespace AvyyanBackend.Controllers
 					StitchLength = productionAllotment.StitchLength,
 					Efficiency = productionAllotment.Efficiency,
 					Composition = productionAllotment.Composition,
-					TotalProductionTime = productionAllotment.TotalProductionTime,
+					TotalProductionTime = (decimal)productionAllotment.TotalProductionTime,
 					CreatedDate = productionAllotment.CreatedDate,
 					YarnLotNo = productionAllotment.YarnLotNo,
 					Counter = productionAllotment.Counter,
@@ -198,7 +198,7 @@ namespace AvyyanBackend.Controllers
 					ReqFinishGsm = productionAllotment.ReqFinishGsm,
 					ReqFinishWidth = productionAllotment.ReqFinishWidth,
 					PartyName = productionAllotment.PartyName,
-					TubeWeight = productionAllotment.TubeWeight,
+					TubeWeight = (decimal)productionAllotment.TubeWeight,
 					TapeColor = productionAllotment.TapeColor
 				};
